Supoort PostgreSQL registry.

+++ bin/genCalibRegistry.py
@@ -17,6 +17,17 @@ import lsst.afw.image  as afwImage
 
 import optparse
 
+try:
+    import psycopg2 as pgsql
+    havePgSql = True
+except ImportError:
+    try:
+        from pg8000 import DBAPI as pgsql
+        havePgSql = True
+    except ImportError:
+        havePgSql = False
+from lsst.daf.butlerUtils import PgSqlConfig
+
 # Needs optparse w/ --create, etc. CPL
 parser = optparse.OptionParser()
 parser.add_option("--create", dest="create", default=False, action="store_true", 
@@ -34,20 +45,42 @@ if len(args) > 0 or len(sys.argv) == 1:
 if opts.camera.lower() not in ("suprime-cam", "suprimecam", "sc", "hsc", "hscsim"):
     raise RuntimeError("Camera not recognised: %s" % camera)
 
-registry = os.path.join(opts.root, "calibRegistry.sqlite3")
+if os.path.exists(os.path.join(opts.root, 'calibRegistry_pgsql.py')) and havePgSql:
+    isSqlite = False
+else:
+    isSqlite = True
 
-if os.path.exists(registry):
-    os.unlink(registry)
-conn = sqlite.connect(registry)
+if isSqlite:
+    registry = os.path.join(opts.root, "calibRegistry.sqlite3")
+
+    if os.path.exists(registry):
+        os.unlink(registry)
+    conn = sqlite.connect(registry)
+else:
+    pgsqlConf = PgSqlConfig()
+    pgsqlConf.load(os.path.join(opts.root, 'calibRegistry_pgsql.py'))
+    conn = pgsql.connect(host=pgsqlConf.host, port=pgsqlConf.port, 
+                         user=pgsqlConf.user, password=pgsqlConf.password,
+                         database=pgsqlConf.db)
+    cur = conn.cursor()
 
 Row = collections.namedtuple("Row", ["calibDate", "calibVersion", "ccd"])
 
 for calib in ('bias', 'dark', 'flat', 'fringe'):
-    cmd = "create table " + calib.lower() + " (id integer primary key autoincrement"
-    cmd += ", validStart text, validEnd text"
-    cmd += ", calibDate text, filter text, calibVersion text, ccd int"
-    cmd += ")"
-    conn.execute(cmd)
+    if isSqlite:
+        cmd = "create table " + calib.lower() + " (id integer primary key autoincrement"
+        cmd += ", validStart text, validEnd text"
+        cmd += ", calibDate text, filter text, calibVersion text, ccd int"
+        cmd += ")"
+        conn.execute(cmd)
+    else:
+        cmd = "DROP TABLE IF EXISTS " + calib.lower()
+        cur.execute(cmd)
+        cmd = "create table " + calib.lower() + " (id SERIAL NOT NULL PRIMARY KEY"
+        cmd += ", validStart VARCHAR(10), validEnd VARCHAR(10)"
+        cmd += ", calibDate VARCHAR(10), filter VARCHAR(16), calibVersion VARCHAR(16), ccd INT"
+        cmd += ")"
+        cur.execute(cmd)
     conn.commit()
 
     rowsPerFilter = dict()
@@ -101,9 +134,15 @@ for calib in ('bias', 'dark', 'flat', 'fringe'):
 
             # print "%f --> %f %f" % (calibDate, validStart, validEnd)
 
-            conn.execute("INSERT INTO " + calib.lower() + " VALUES (NULL, ?, ?, ?, ?, ?, ?)",
-                         (validStart, validEnd, calibDate, filterName, row.calibVersion, row.ccd))
+            if isSqlite:
+                conn.execute("INSERT INTO " + calib.lower() + " VALUES (NULL, ?, ?, ?, ?, ?, ?)",
+                             (validStart, validEnd, calibDate, filterName, row.calibVersion, row.ccd))
+            else:
+                cur.execute("INSERT INTO " + calib.lower() + " (validStart, validEnd, calibDate, filter, calibVersion, ccd) VALUES (%s, %s, %s, %s, %s, %s)",
+                            (validStart, validEnd, calibDate, filterName, row.calibVersion, row.ccd))
 
 
 conn.commit()
+if not isSqlite:
+    cur.close()
 conn.close()
